Refactor UI components
